`g resource` may use @Patch() in favor of @Put()
Using `generate resource` creates an `update()` function annotated with `@Put()`. As of my understanding of REST API's `PUT` is meant to replace a full object or create one if it is not existing. But the created function clearly updates only specific values and integrates validation by use of `PartialType()`. Therefor, imo, `@Patch()` should be used instead which is meant to update only specific values of an existing object.